dch-2 refactor web-publisher-serv
