+++ web-publisher-service/src/main/java/com/devcomhub/webpublisherservice/dto/PaymentSendRequest.java
@@ -5,10 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Builder
 @Data
-@NoArgsConstructor
-@AllArgsConstructor
 public class PaymentSendRequest {
     private String paymentId;
     private Double amount;
