+++ web-publisher-service/src/main/resources/application.yaml
@@ -1,5 +1,5 @@
 server:
-  port: 8082
+  port: 8080
 
 devcomhub:
   kafka-prop:
